requirements.txt
discord.py>=2.0
yt-dlp
python-dotenv
pynacl
